requirements.txt
anthropic
tree-sitter
tree-sitter-c
python-dotenv